Repackaged into legacy folder. Moved handler into lib. Updated references.
