+++ src/lib/legacy/deterministic-hazard-factory.js
