+++ src/lib/legacy/legacy-factory.js
