+++ src/lib/legacy/metadata-factory.js
