+++ src/lib/legacy/probabilistic-hazard-factory.js
